refactor: update all references to use MaterialModal
- Update imports to use MaterialModal instead of MaterialSelector/MaterialCenterModal
- Home.tsx: use MaterialModal mode='manage'
- SlidePreview.tsx: use MaterialModal mode='select'
- MaterialGeneratorModal.tsx: use MaterialModal mode='select'
- TemplateSelector.tsx: use MaterialModal mode='select'
- Update index.ts to export MaterialModal and materialUrlToFile
