+++ frontend/src/components/shared/index.ts
@@ -9,12 +9,11 @@ export { StatusBadge } from './StatusBadge';
 export { ContextualStatusBadge } from './ContextualStatusBadge';
 export { ConfirmDialog, useConfirm } from './ConfirmDialog';
 export { MaterialGeneratorModal } from './MaterialGeneratorModal';
-export { MaterialCenterModal } from './MaterialCenterModal';
+export { MaterialModal, materialUrlToFile } from './MaterialModal';
 export { ReferenceFileCard } from './ReferenceFileCard';
 export { ReferenceFileSelector } from './ReferenceFileSelector';
 export { FilePreviewModal } from './FilePreviewModal';
 export { ReferenceFileList } from './ReferenceFileList';
-export { MaterialSelector } from './MaterialSelector';
 export { Footer } from './Footer';
 export { GithubBadge } from './GithubBadge';
 export { GithubRepoCard } from './GithubRepoCard';
